Close #1, finish inner meesage feature

+++ src/main/java/com/qxcmp/framework/web/view/elements/menu/item/BackendAccountAlarmItem.java
@@ -10,9 +10,9 @@ import lombok.Getter;
 @Getter
 public class BackendAccountAlarmItem extends AbstractMenuItem {
 
-    private int messageCount;
+    private long messageCount;
 
-    public BackendAccountAlarmItem(int messageCount) {
+    public BackendAccountAlarmItem(long messageCount) {
         this.messageCount = messageCount;
     }
